Split the existing gulpfile into seperate modules
Modules now live in lib/gulp. Each module requires just the code it needs to work. Rather than use a requireDir solution the modules are individually required at the top of gulpfile.js. I’ve left the top level tasks in the main gulpfile for clarity of what’s available, but I could be argued into moving these into their respective modules if people feel that it’s a better idea. Single-command tasks I’ve left in the main gulpfile too - again, happy to change this if people want.
